Try out a new interface for Umem and Frames
This should make it easier to handle Frames living in Umem similar
to a Box<[u8]>. Frames now own the MmapArea in which they live via
a shared Arc.

This is not yet finished as all the tests and dev1_to_dev2 examples
still need to be converted to the new interface. But it should give
an a idea of how a new interface could look like.